Require IMAGE_NAME in all tests, do not default to public image

+++ images/opensearch/tests/01-version.sh
@@ -2,7 +2,9 @@
 
 set -o errexit -o nounset -o errtrace -o pipefail -x
 
-IMAGE_DIR="$(basename "$(cd "$(dirname ${BASH_SOURCE[0]})/.." && pwd )")"
-IMAGE_NAME=${IMAGE_NAME:-"cgr.dev/chainguard/${IMAGE_DIR}:latest"}
+if [[ "${IMAGE_NAME}" == "" ]]; then
+    echo "Must set IMAGE_NAME environment variable. Exiting."
+    exit 1
+fi
 
 docker run --rm --entrypoint=/usr/bin/opensearch "${IMAGE_NAME}" --version
